Allow unpacking bit-packed data without NumPy
