+++ doc/release_notes/v2.3.0.rst
@@ -24,6 +24,8 @@ Enhancements
 * Use rle_handler as last resort handler for decoding RLE encoded data as it is
   the slowest handler (:issue:`1487`)
 * Added, enhanced, or removed a number of Mitra private dictionary entries (:pr:`1588`)
+* Added support for unpacking bit-packed data without using NumPy to
+  :func:`~pydicom.pixel_data_handlers.utils.unpack_bits`(:pr:`1594`)
 
 
 Fixes
