+++ pydicom/pixel_data_handlers/__init__.py
@@ -1,7 +1,13 @@
 
 from pydicom.pixel_data_handlers.util import (
-    apply_color_lut, apply_modality_lut, apply_voi_lut, convert_color_space,
-    apply_voi, apply_windowing
+    apply_color_lut,
+    apply_modality_lut,
+    apply_voi_lut,
+    convert_color_space,
+    apply_voi,
+    apply_windowing,
+    pack_bits,
+    unpack_bits,
 )
 
 apply_rescale = apply_modality_lut
